refactor: 추상화를 통한 Repository 패턴 적용
