+++ src/main/java/cart/application/OrderService.java
@@ -1,6 +1,5 @@
 package cart.application;
 
-import cart.dao.ProductDao;
 import cart.domain.*;
 import cart.dto.*;
 import cart.exception.OrderException;
