+++ src/test/java/cart/integration/OrderIntegrationTest.java
@@ -1,6 +1,5 @@
 package cart.integration;
 
-import cart.dao.MemberDao;
 import cart.domain.Member;
 import cart.dto.OrderRequest;
 import cart.dto.ProductOrderRequest;
