+++ src/test/java/cart/repository/OrderRepositoryTest.java
@@ -8,7 +8,6 @@ import cart.entity.OrderEntity;
 import cart.entity.OrderItemEntity;
 import cart.entity.PointHistoryEntity;
 import cart.exception.OrderException;
-import cart.exception.OrderServerException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
